Refactor unsupervised metrics into a separate module
Moved `UnsupervisedMetric` and `UnsupervisedNumpyMetric` from `base_metrics.py` to a new `unsupervised_metrics.py` module. This improves modularity and separation of concerns within the codebase. Adjusted imports in `__init__.py` accordingly to reflect the changes.
