+++ pytorch_tabnet/metrics/__init__.py
@@ -13,12 +13,6 @@ from .base_metrics import (
     UnsupervisedLoss as UnsupervisedLoss,
 )
 from .base_metrics import (
-    UnsupervisedMetric as UnsupervisedMetric,
-)
-from .base_metrics import (
-    UnsupervisedNumpyMetric as UnsupervisedNumpyMetric,
-)
-from .base_metrics import (
     UnsupMetricContainer as UnsupMetricContainer,
 )
 from .base_metrics import (
@@ -29,3 +23,9 @@ from .mae import MAE as MAE
 from .mse import MSE as MSE
 from .rmse import RMSE as RMSE
 from .rmsle import RMSLE as RMSLE
+from .unsupervised_metrics import (
+    UnsupervisedMetric as UnsupervisedMetric,
+)
+from .unsupervised_metrics import (
+    UnsupervisedNumpyMetric as UnsupervisedNumpyMetric,
+)
