+++ pytorch_tabnet/metrics/base_metrics.py
@@ -192,62 +192,3 @@ def check_metrics(metrics: List[Union[str, Any]]) -> List[str]:
         else:
             raise TypeError("You need to provide a valid metric format")
     return val_metrics
-
-
-class UnsupervisedMetric(Metric):
-    """Unsupervised loss metric for reconstruction tasks."""
-
-    _name: str = "unsup_loss"
-    _maximize: bool = False
-
-    def __call__(self, y_pred: torch.Tensor, embedded_x: torch.Tensor, obf_vars: torch.Tensor, weights: torch.Tensor = None) -> float:  # type: ignore[override]
-        """Compute the unsupervised loss metric.
-
-        Parameters
-        ----------
-        y_pred : torch.Tensor
-            Predicted values.
-        embedded_x : torch.Tensor
-            Embedded input values.
-        obf_vars : torch.Tensor
-            Obfuscated variables mask.
-        weights : torch.Tensor, optional
-            Sample weights.
-
-        Returns
-        -------
-        float
-            The computed unsupervised loss metric.
-
-        """
-        loss = UnsupervisedLoss(y_pred, embedded_x, obf_vars, weights=weights)
-        return loss.cpu().item()
-
-
-class UnsupervisedNumpyMetric(Metric):
-    """Unsupervised loss metric (NumPy version) for reconstruction tasks."""
-
-    _name: str = "unsup_loss_numpy"
-    _maximize: bool = False
-
-    def __call__(self, y_pred: torch.Tensor, embedded_x: torch.Tensor, obf_vars: torch.Tensor, weights: torch.Tensor = None) -> float:  # type: ignore[override]
-        """Compute the unsupervised loss metric (NumPy version).
-
-        Parameters
-        ----------
-        y_pred : torch.Tensor
-            Predicted values.
-        embedded_x : torch.Tensor
-            Embedded input values.
-        obf_vars : torch.Tensor
-            Obfuscated variables mask.
-        weights : torch.Tensor, optional
-            Sample weights.
-
-        Returns
-        -------
-        float
-            The computed unsupervised loss metric (NumPy version).
-
-        """
-        return UnsupervisedLoss(y_pred, embedded_x, obf_vars).cpu().item()
